Add Components, move files into detail
