+++ include/goofit/PDFs/Component.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <goofit/PDFs/GooPdf.h>
+
+namespace GooFit {
+
+class Component : public GooPdf {
+  public:
+    using GooPdf::GooPdf;
+};
+
+} // namespace GooFit
