+++ include/goofit/PDFs/physics/Amp3Body_TD.h
@@ -72,26 +72,26 @@ with a working example and modify it gradually.
 class Amp3Body_TD : public Amp3BodyBase {
   public:
     Amp3Body_TD(std::string n,
-              Observable _dtime,
-              Observable _sigmat,
-              Observable m12,
-              Observable m13,
-              EventNumber eventNumber,
-              DecayInfo3t decay,
-              MixingTimeResolution *r,
-              GooPdf *eff,
-              Observable *mistag = nullptr);
+                Observable _dtime,
+                Observable _sigmat,
+                Observable m12,
+                Observable m13,
+                EventNumber eventNumber,
+                DecayInfo3t decay,
+                MixingTimeResolution *r,
+                GooPdf *eff,
+                Observable *mistag = nullptr);
     Amp3Body_TD(std::string n,
-              Observable _dtime,
-              Observable _sigmat,
-              Observable m12,
-              Observable m13,
-              EventNumber eventNumber,
-              DecayInfo3t decay,
-              std::vector<MixingTimeResolution *> &r,
-              GooPdf *eff,
-              Observable md0,
-              Observable *mistag = nullptr);
+                Observable _dtime,
+                Observable _sigmat,
+                Observable m12,
+                Observable m13,
+                EventNumber eventNumber,
+                DecayInfo3t decay,
+                std::vector<MixingTimeResolution *> &r,
+                GooPdf *eff,
+                Observable md0,
+                Observable *mistag = nullptr);
     // Note that 'efficiency' refers to anything which depends on (m12, m13) and multiplies the
     // coherent sum. The caching method requires that it be done this way or the ProdPdf
     // normalization will get *really* confused and give wrong answers.
