+++ include/goofit/PDFs/physics/AmpComponent.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <goofit/PDFs/ComplexComponent.h>
+
+namespace GooFit {
+
+class AmpComponent : public ComplexComponent {
+  public:
+    using ComplexComponent::ComplexComponent;
+};
+
+} // namespace GooFit
