+++ python/PDFs/physics/ThreeGaussResolution.cpp
