+++ src/PDFs/physics/SpinFactors.cu
@@ -520,7 +520,7 @@ __device__ spin_function_ptr ptr_to_ONE         = ONE;
 
 SpinFactor::SpinFactor(
     std::string name, SF_4Body SF, fptype mD0, unsigned int P0, unsigned int P1, unsigned int P2, unsigned int P3)
-    : GooPdf(name)
+    : AmpComponent(name)
     , _SF(SF)
     , _P0(P0)
     , _P1(P1)
