+++ src/PDFs/physics/detail/AmpCalc.cu
@@ -1,26 +1,26 @@
-#include <goofit/PDFs/physics/detail/AmpCalc_TD.h>
+#include <goofit/PDFs/physics/detail/AmpCalc.h>
 
 #include <goofit/PDFs/ParameterContainer.h>
 #include <goofit/PDFs/physics/Amp4BodyGlobals.h>
-#include <goofit/PDFs/physics/detail/EvalVar.h>
 #include <goofit/PDFs/physics/SpinFactors.h>
+#include <goofit/PDFs/physics/detail/EvalVar.h>
 
 namespace GooFit {
 
-AmpCalc_TD::AmpCalc_TD(unsigned int nPerm, unsigned int ampIdx)
+AmpCalc::AmpCalc(unsigned int nPerm, unsigned int amp)
     : _nPerm(nPerm)
-    , _AmpIdx(ampIdx) {}
+    , _AmpIdx(amp) {}
 
-__device__ fpcomplex AmpCalc_TD::operator()(thrust::tuple<int, fptype *, int> t) const {
+__device__ fpcomplex AmpCalc::operator()(thrust::tuple<int, fptype *, int> t) const {
     ParameterContainer pc;
 
     while(pc.funcIdx < dalitzFuncId)
         pc.incrementIndex();
 
-    unsigned int cacheToUse = pc.getConstant(6);
-    unsigned int totalLS    = pc.getConstant(7);
-    unsigned int totalSF    = pc.getConstant(8);
-    unsigned int totalAMP   = pc.getConstant(9);
+    unsigned int cacheToUse = pc.getConstant(5);
+    unsigned int totalLS    = pc.getConstant(6);
+    unsigned int totalSF    = pc.getConstant(7);
+    unsigned int totalAMP   = pc.getConstant(8);
     unsigned int offset     = totalLS + totalSF;
     unsigned int numLS      = AmpIndices[totalAMP + _AmpIdx];
     unsigned int numSF      = AmpIndices[totalAMP + _AmpIdx + 1];
@@ -32,21 +32,21 @@ __device__ fpcomplex AmpCalc_TD::operator()(thrust::tuple<int, fptype *, int> t)
 
     for(int i = 0; i < _nPerm; ++i) {
         fpcomplex ret(1, 0);
+        fpcomplex tmp(1, 0);
 
         for(int j = i * LS_step; j < (i + 1) * LS_step; ++j) {
-            int idx = AmpIndices[totalAMP + _AmpIdx + 4 + j];
-            ret *= (cResSF_TD[cacheToUse][evtNum * offset + idx]);
-            // printf("Lineshape %i = (%.7g, %.7g)\n", j, (cResSF_TD[cacheToUse][evtNum*offset + AmpIndices[totalAMP +
-            // _AmpIdx + 4 + j]]).real, (cResSF_TD[cacheToUse][evtNum*offset + AmpIndices[totalAMP + _AmpIdx + 4 +
-            // j]]).imag);
+            int idx = AmpIndices[totalAMP + _AmpIdx + 3 + j];
+            tmp     = (cResSF[cacheToUse][evtNum * offset + idx]);
+            ret *= tmp;
+            // printf("Lineshape = (%.7g, %.7g)\n", tmp.real, tmp.imag);
         }
 
         // printf("Lineshape Product = (%.7g, %.7g)\n", ret.real, ret.imag);
         for(int j = i * SF_step; j < (i + 1) * SF_step; ++j) {
-            int idx = AmpIndices[totalAMP + _AmpIdx + 4 + numLS + j];
-            ret *= (cResSF_TD[cacheToUse][evtNum * offset + totalLS + idx].real());
-            // printf(" SF = %.7g\n", (cResSF_TD[cacheToUse][evtNum*offset + totalLS + AmpIndices[totalAMP + _AmpIdx + 4
-            // + numLS + j]].real));
+            int idx = AmpIndices[totalAMP + _AmpIdx + 3 + numLS + j];
+            tmp     = (cResSF[cacheToUse][evtNum * offset + totalLS + idx].real());
+            ret *= tmp;
+            // printf("SF = (%.7g, %.7g)\n", tmp.real, tmp.imag);
         }
 
         // printf("Lineshape Product * SF = (%.7g, %.7g)\n", ret.real, ret.imag);
