+++ src/PDFs/physics/detail/NormLSCalculator.cu
@@ -1,37 +1,29 @@
-#include <goofit/PDFs/physics/detail/LSCalculator.h>
+#include <goofit/PDFs/physics/detail/NormLSCalculator.h>
 
 #include <goofit/PDFs/ParameterContainer.h>
-#include <goofit/PDFs/physics/detail/EvalVar.h>
 #include <goofit/PDFs/physics/SpinFactors.h>
+#include <goofit/PDFs/physics/detail/EvalVar.h>
 
 namespace GooFit {
 
-LSCalculator::LSCalculator() = default;
+NormLSCalculator::NormLSCalculator() = default;
 
-__device__ fpcomplex LSCalculator::operator()(thrust::tuple<int, fptype *, int> t) const {
+__device__ fpcomplex NormLSCalculator::operator()(
+    thrust::tuple<mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t, mcbooster::GReal_t> t)
+    const {
     // Calculates the BW values for a specific resonance.
     fpcomplex ret;
 
-    int evtNum  = thrust::get<0>(t);
-    fptype *evt = thrust::get<1>(t) + (evtNum * thrust::get<2>(t));
-
     ParameterContainer pc;
 
-    // Increment to DP
     while(pc.funcIdx < dalitzFuncId)
         pc.incrementIndex();
 
-    int id_m12   = pc.getObservable(0);
-    int id_m34   = pc.getObservable(1);
-    int id_cos12 = pc.getObservable(2);
-    int id_cos34 = pc.getObservable(3);
-    int id_phi   = pc.getObservable(4);
-
-    fptype m12   = evt[id_m12];
-    fptype m34   = evt[id_m34];
-    fptype cos12 = evt[id_cos12];
-    fptype cos34 = evt[id_cos34];
-    fptype phi   = evt[id_phi];
+    fptype m12   = (thrust::get<0>(t));
+    fptype m34   = (thrust::get<1>(t));
+    fptype cos12 = (thrust::get<2>(t));
+    fptype cos34 = (thrust::get<3>(t));
+    fptype phi   = (thrust::get<4>(t));
 
     fptype M  = pc.getConstant(0);
     fptype m1 = pc.getConstant(1);
@@ -39,6 +31,7 @@ __device__ fpcomplex LSCalculator::operator()(thrust::tuple<int, fptype *, int> 
     fptype m3 = pc.getConstant(3);
     fptype m4 = pc.getConstant(4);
 
+    // skip to our resonance function
     while(pc.funcIdx < _resonance_i)
         pc.incrementIndex();
 
@@ -49,21 +42,21 @@ __device__ fpcomplex LSCalculator::operator()(thrust::tuple<int, fptype *, int> 
         fptype d1   = pair == 0 ? m1 : m3;
         fptype d2   = pair == 0 ? m2 : m4;
         ret         = getResonanceAmplitude(mres, d1, d2, pc);
-        // printf("LS %i: mass:%f, %f i%f\n",_resonance_i, mres, ret.real, ret.imag );
     } else {
         fptype vecs[16];
+        // TODO: What is indices[1]?
         get4Vecs(vecs, m12, m34, cos12, cos34, phi, M, m1, m2, m3, m4);
         fptype d1, d2;
         fptype mres = getmass(pair, d1, d2, vecs, m1, m2, m3, m4);
         ret         = getResonanceAmplitude(mres, d1, d2, pc);
-        // printf("LS_m_calc %i: mass:%f, %f i%f\n",_resonance_i, mres, ret.real, ret.imag );
     }
 
-    // if (!inDalitz(m12, m13, motherMass, daug1Mass, daug2Mass, daug3Mass)) return ret;
+    // printf("NormLS %f, %f, %f, %f, %f \n",m12, m34, cos12, cos34, phi );
+    // printf("%i, %i, %i, %i, %i \n",numLS, numSF, numAmps, offset, evtNum );
+    // printf("NLS %i, %f, %f\n",_resonance_i,ret.real, ret.imag);
+
     // printf("m12 %f \n", m12); // %f %f %f (%f, %f)\n ", m12, m13, m23, ret.real, ret.imag);
     // printf("#Parameters %i, #LS %i, #SF %i, #AMP %i \n", indices[0], indices[3], indices[4], indices[5]);
-    // printf("%i mass: %.5g, BW_%i : %f %f\n",evtNum, massstore, _resonance_i, ret.real, ret.imag);
-
     return ret;
 }
