+++ src/PDFs/physics/detail/SpecialDalitzIntegrator.cu
