+++ src/PDFs/physics/detail/SpecialResonanceCalculator.cu
