+++ src/PDFs/physics/detail/SpecialResonanceIntegrator.cu
