+++ src/PDFs/physics/detail/SpecialWaveCalculator.cu
